backend/requirements.txt
Flask==3.0.0
Pillow==10.0.0
flask-cors==4.0.0
